# backend/openapi.yaml
{
  "openapi": "3.0.3",
  "info": {
    "title": "SocialFlow AI Dashboard API",
    "version": "1.0.0",
    "description": "REST API for the SocialFlow AI Dashboard — social media scheduling, analytics, AI content tools, billing, and more."
  },
  "servers": [
    { "url": "/api/v1", "description": "Current stable version" },
    { "url": "/api",    "description": "Legacy alias (deprecated)" }
  ],
  "security": [{ "bearerAuth": [] }],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "WebhookSubscription": {
        "type": "object",
        "properties": {
          "id":        { "type": "string", "format": "uuid" },
          "url":       { "type": "string", "format": "uri" },
          "events":    { "type": "array", "items": { "type": "string" } },
          "isActive":  { "type": "boolean" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "Subscription": {
        "type": "object",
        "properties": {
          "id":      { "type": "string" },
          "userId":  { "type": "string" },
          "plan":    { "type": "string" },
          "credits": { "type": "number" },
          "status":  { "type": "string" }
        }
      },
...
    }
  },
  "paths": {
...
    "/billing/provision": {
      "post": {
        "tags": ["Billing"],
        "summary": "Provision a Stripe customer and free subscription for the authenticated user",
        "responses": {
          "201": { "description": "Subscription provisioned", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Subscription" } } } },
          "404": { "description": "User not found" },
          "502": { "description": "Stripe error" }
        }
      }
    },
    "/billing/subscription": {
      "get": {
        "tags": ["Billing"],
        "summary": "Get the current user's subscription and credit balance",
        "responses": {
          "200": { "description": "Subscription details", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Subscription" } } } },
          "404": { "description": "No subscription found" }
        }
      }
    },
...
    "/billing/checkout": {
      "post": {
        "tags": ["Billing"],
        "summary": "Create a Stripe Checkout session for upgrading to a paid plan",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["priceId","successUrl","cancelUrl"],
                "properties": {
                  "priceId":    { "type": "string" },
                  "successUrl": { "type": "string", "format": "uri" },
                  "cancelUrl":  { "type": "string", "format": "uri" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Checkout session URL", "content": { "application/json": { "schema": { "type": "object", "properties": { "url": { "type": "string", "format": "uri" } } } } } },
          "502": { "description": "Stripe error" }
        }
      }
    },
    "/billing/portal": {
      "post": {
        "tags": ["Billing"],
        "summary": "Create a Stripe Customer Portal session",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["returnUrl"],
                "properties": { "returnUrl": { "type": "string", "format": "uri" } }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Portal session URL", "content": { "application/json": { "schema": { "type": "object", "properties": { "url": { "type": "string", "format": "uri" } } } } } },
          "502": { "description": "Stripe error" }
        }
      }
    },
    "/billing/webhook": {
      "post": {
        "tags": ["Billing"],
        "summary": "Stripe webhook receiver (raw body, HMAC-verified)",
        "security": [],
        "parameters": [
          { "in": "header", "name": "stripe-signature", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/octet-stream": { "schema": { "type": "string", "format": "binary" } } }
        },
        "responses": {
          "200": { "description": "Event received" },
          "400": { "description": "Missing signature or invalid payload" }
        }
      }
    },
    "/webhooks": {
      "get": {
        "tags": ["Webhooks"],
        "summary": "List all webhook subscriptions for the authenticated user",
        "responses": {
          "200": { "description": "Array of webhook subscriptions", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/WebhookSubscription" } } } } },
          "401": { "description": "Unauthorized" }
        }
      },
      "post": {
        "tags": ["Webhooks"],
        "summary": "Create a new webhook subscription",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["url","secret","events"],
                "properties": {
                  "url":    { "type": "string", "format": "uri" },
                  "secret": { "type": "string", "minLength": 16 },
                  "events": { "type": "array", "items": { "type": "string" }, "minItems": 1 }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Webhook created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/WebhookSubscription" } } } },
          "400": { "description": "Validation error" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/webhooks/{id}": {
      "get": {
        "tags": ["Webhooks"],
        "summary": "Get a webhook subscription by ID",
        "parameters": [{ "in": "path", "name": "id", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": {
          "200": { "description": "Webhook subscription", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/WebhookSubscription" } } } },
          "404": { "description": "Not found" }
        }
      },
      "patch": {
        "tags": ["Webhooks"],
        "summary": "Update a webhook subscription",
        "parameters": [{ "in": "path", "name": "id", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "url":      { "type": "string", "format": "uri" },
                  "secret":   { "type": "string", "minLength": 16 },
                  "events":   { "type": "array", "items": { "type": "string" } },
                  "isActive": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Updated webhook" },
          "404": { "description": "Not found" }
        }
      },
      "delete": {
        "tags": ["Webhooks"],
        "summary": "Delete a webhook subscription",
        "parameters": [{ "in": "path", "name": "id", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": {
          "204": { "description": "Deleted" },
          "404": { "description": "Not found" }
        }
      }
    },
...
  }
}